# ledgerx-sdk/openapi.yaml
openapi: 3.0.3
info:
  title: LedgerX Public API
  version: 1.0.0
  description: |
    ## Introduction

    The LedgerX API provides endpoints for synchronizing on-chain wallets, retrieving
    normalized financial transactions, and managing wallets, wallet groups, organization
    members, and the accounting system (Chart of Accounts).

    - **Security**: All endpoints (unless otherwise specified) require an **API_KEY**
      in the request header.
    - **Rate Limit**: Depends on the user’s subscription plan. See details on the
      subscriptions page.
    - **Compute Unit (CU)**: Each API consumes a specific number of CUs, which are
      deducted from your monthly total.

    **Useful Links**:
    - Rate limit: https://ledgerx.finance/subscriptions
    - Compute Unit: https://ledgerx.finance/subscriptions

    > Note: Some CU and rate limit descriptions are for documentation purposes only.
    Actual limits are applied based on the user’s subscription plan.
servers:
  - url: https://api.ledgerx.finance
    description: Production
  - url: https://sandbox.api.ledgerx.finance
    description: Sandbox
security:
  - ApiKeyAuth: []
...
paths:
...
  /chart-of-accounts:
    get:
      tags: [Chart of Accounts]
      summary: List Chart of Accounts
      operationId: listChartOfAccounts
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        code:
                          type: string
                        type:
                          type: string
                        description:
                          type: string
                        status:
                          type: string
    post:
      tags: [Chart of Accounts]
      summary: Create a Chart of Account
      operationId: createChartOfAccount
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, type]
              properties:
                description:
                  type: string
                code:
                  type: string
                name:
                  type: string
                type:
                  type: string
                  enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
              example:
                description: ''
                code: '106'
                name: Investment bonding
                type: ASSET
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      code:
                        type: string
                      type:
                        type: string
                      description:
                        type: string
                      status:
                        type: string
  /chart-of-accounts/{coaId}:
    put:
      tags: [Chart of Accounts]
      summary: Update a Chart of Account
      operationId: updateChartOfAccount
      security:
        - ApiKeyAuth: []
      parameters:
        - name: coaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, type]
              properties:
                code:
                  type: string
                name:
                  type: string
                type:
                  type: string
                  enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
                description:
                  type: string
              example:
                code: '106'
                name: Investment bonding 123
                type: ASSET
                description: ''
      responses:
        '204':
          description: No Content
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: API_KEY
      description: API key created in **User Dashboard > API**.
...